+++ asf_search/Products/NISARProduct.py
@@ -1,4 +1,4 @@
-from typing import Dict, Union
+from typing import Dict, Tuple, Union
 from asf_search import ASFSearchOptions, ASFSession, ASFStackableProduct
 from asf_search.CMR.translate import try_parse_float, try_parse_int, try_round_float
 from asf_search.constants import PRODUCT_TYPE
@@ -11,6 +11,7 @@ class NISARProduct(ASFStackableProduct):
     ASF Dataset Documentation Page: https://asf.alaska.edu/nisar/
     """
     _base_properties = {
+        **ASFStackableProduct._base_properties,
         'pgeVersion': {'path': ['PGEVersionClass', 'PGEVersion']}
     }
 
@@ -40,18 +41,11 @@ class NISARProduct(ASFStackableProduct):
         :return: ASFSearchOptions describing appropriate options for building a stack from this product
         """
         return None
-    
-    @staticmethod
-    def get_property_paths() -> Dict:
-        return {
-            **ASFStackableProduct.get_property_paths(),
-            **NISARProduct._base_properties
-        }
 
-    def get_sort_keys(self):
+    def get_sort_keys(self) -> Tuple[str, str]:
         keys = super().get_sort_keys()
-
-        if keys[0] is None:
-            return (self.properties.get('processingDate', ''), keys[1])
+        
+        if keys[0] == '':
+            return (self._read_property('processingDate', ''), keys[1])
 
         return keys