+++ asf_search/Products/OPERAS1Product.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import Dict, Tuple
 from asf_search import ASFSearchOptions, ASFSession
 from asf_search.CMR.translate import try_parse_date
 from asf_search.Products import S1Product
@@ -9,6 +9,7 @@ class OPERAS1Product(S1Product):
     ASF Dataset Documentation Page: https://asf.alaska.edu/datasets/daac/opera/
     """
     _base_properties = {
+        **S1Product._base_properties,
         'centerLat': {'path': []}, # Opera products lacks these fields
         'centerLon': {'path': []},
         'frameNumber': {'path': []},
@@ -18,6 +19,8 @@ class OPERAS1Product(S1Product):
         'subswath': {'path': ['AdditionalAttributes', ('Name', 'SUBSWATH_NAME'), 'Values', 0]},
         'polarization': {'path': ['AdditionalAttributes', ('Name', 'POLARIZATION'), 'Values']} # dual polarization is in list rather than a 'VV+VH' style format
     }
+
+    _subclass_concept_ids = { 'C1257995185-ASF', 'C1257995186-ASF', 'C1258354200-ASF', 'C1258354201-ASF', 'C1259974840-ASF', 'C1259976861-ASF', 'C1259981910-ASF', 'C1259982010-ASF', 'C2777436413-ASF', 'C2777443834-ASF', 'C2795135174-ASF', 'C2795135668-ASF','C1260721853-ASF', 'C1260721945-ASF', 'C2803501097-ASF', 'C2803501758-ASF' }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
@@ -47,13 +50,6 @@ class OPERAS1Product(S1Product):
         return opts
 
     @staticmethod
-    def get_property_paths() -> Dict:
-        return {
-            **S1Product.get_property_paths(),
-            **OPERAS1Product._base_properties
-        }
-    
-    @staticmethod
     def get_default_baseline_product_type() -> None:
         """
         Returns the product type to search for when building a baseline stack.
@@ -71,10 +67,17 @@ class OPERAS1Product(S1Product):
         """
         return None
 
-    def get_sort_keys(self):
+    def get_sort_keys(self) -> Tuple[str, str]:
         keys = super().get_sort_keys()
 
-        if keys[0] is None:
-            keys = self.properties.get('validityStartDate'), keys[1]
+        if keys[0] == '':
+            return (self._read_property('validityStartDate', ''), keys[1])
 
         return keys
+
+    @staticmethod
+    def _is_subclass(item: Dict) -> bool:
+        # not all umm products have this field set, 
+        # but when it's available it's convenient for fast matching
+        concept_id = item['meta'].get('collection-concept-id')
+        return concept_id in OPERAS1Product._subclass_concept_ids