+++ tests/ASFSession/test_ASFSession.py
@@ -16,7 +16,7 @@ def run_auth_with_creds(username: str, password: str):
 def run_auth_with_token(token: str):
     session = ASFSession()
 
-    with patch('asf_search.ASFSession.get') as mock_token_session:
+    with patch('asf_search.ASFSession.post') as mock_token_session:
         if not token.startswith('Bearer EDL'):
                 mock_token_session.return_value.status_code = 400
                 session.auth_with_token(token)
@@ -28,8 +28,13 @@ def run_auth_with_cookiejar(cookies: List):
     cookiejar = http.cookiejar.CookieJar()
     for cookie in cookies:
         cookiejar.set_cookie(create_cookie(name=cookie.pop('name'), **cookie))
+
+    # requests.cookies.RequestsCookieJar, which has slightly different behaviour
     session = ASFSession()
-    session.auth_with_cookiejar(cookies)
+    session.auth_with_cookiejar(cookiejar)
+
+    request_cookiejar_session = ASFSession()
+    request_cookiejar_session.auth_with_cookiejar(session.cookies)
 
 def run_test_asf_session_rebuild_auth(
     original_domain: str, 
@@ -43,7 +48,7 @@ def run_test_asf_session_rebuild_auth(
 
     session = ASFSession()
 
-    with patch('asf_search.ASFSession.get') as mock_token_session:
+    with patch('asf_search.ASFSession.post') as mock_token_session:
         mock_token_session.return_value.status_code = 200
         session.auth_with_token("bad_token")